Extract DeletionVector logic from PuffinFile
